UART header source
UART functionality is moved to UART source file
